--- src/app/consts/platform.ts
@@ -1,11 +1,0 @@
-export enum wmPlatformType {
-  Mac = "mac",
-  Mock = "mock",
-  Web = "web",
-  Windows = "windows",
-}
-
-export enum tmPlatformType {
-  App = "app",
-  Web = "web",
-}
